moved db into medusa folder, training and scorer uses MedusaMongo
