+++ Medusa/Database/MoveBackTrainingScripts.txt
