+++ Medusa/MedusaMongo.py
@@ -4,6 +4,7 @@ import pymongo
 
 uri = "mongodb://MedusaUser:P3R5EU?@applis.me/Medusa?authSource=Medusa"
 
+# Used in Medusa - Training
 def getMedusaTrainingCollection():
     mongoClient=pymongo.MongoClient(uri)
     medusaDB = mongoClient["Medusa"]
@@ -26,3 +27,12 @@ def get_next_n_samples(cursor,n = 100):
         entry = cursor.next()
         samples.append((entry["success"],entry["image"]))
     return samples
+
+# Used in Scorer
+def save_results_to_Mongo(scores,imgBytes):
+    imageCollection.insert_one({
+        "scores":scores,
+        "insert_date":str(datetime.datetime.now()),
+        "image":imgBytes
+    }
+)
